Factor out ipopt_solver_internal.
